fix(jco): address code review
